MERGE: Clear session-only interest groups on session close
- Change Interest Group storage to accept an origin matcher instead of an origin
- Add StoragePartition REMOVE_DATA_MASK for interest groups and use that to control clearing interest groups.
- Set interest group storage to be cleared where cookies are currently cleared

(cherry picked from commit 3ff626734f85314c943e9fab88fcf11d76a29005)

Bug: 1199852, 1202133
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2844632
Commit-Queue: Russ Hamilton <behamilton@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#876140}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2862286
Cr-Commit-Position: refs/branch-heads/4472@{#677}
Cr-Branched-From: 3d60439cfb36485e76a1c5bb7f513d3721b20da1-refs/heads/master@{#870763}

+++ content/browser/storage_partition_impl.cc
@@ -2225,10 +2225,13 @@ void StoragePartitionImpl::DataDeletionHelper::ClearDataOnUIThread(
             // being called.
             mojo::WrapCallbackWithDefaultInvokeIfNotRun(
                 CreateTaskCompletionClosure(TracingDataType::kCookies))));
+  }
 
+  if (remove_mask_ & REMOVE_DATA_MASK_INTEREST_GROUPS) {
     if (interest_group_manager) {
       interest_group_manager->DeleteInterestGroupData(
-          url::Origin::Create(storage_origin));
+          CreateGenericOriginMatcher(storage_origin, origin_matcher,
+                                     storage_policy_ref));
     }
   }
